requirements.txt
numpy==1.22.0
scikit-learn==0.24.1
scipy==1.11.1
pyyaml==6.0
tqdm==4.65.0
-f https://download.pytorch.org/whl/torch_stable.html 
torch==1.11.0+cu113 
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.12.0+cu113
-f https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.11.0+cu113
